WIP: refactor scene management.
