+++ app/development/contextMenu/setState.ts
@@ -1,4 +1,3 @@
-import {applySavedState, getState} from 'app/state';
 import {setSpawnLocation} from 'app/content/spawnLocations';
 import {
     earlyDungeonSpawnLocations,
@@ -9,6 +8,8 @@ import {
     devSpawnLocations,
     SpawnLocationOptions,
 } from 'app/content/spawnStates';
+import {applySavedState} from 'app/scenes/fileSelect/setSaveFileToState';
+import {getState} from 'app/state';
 import {cloneDeep} from 'app/utils/index';
 import {returnToSpawnLocation} from 'app/utils/returnToSpawnLocation';
 import {setSaveSlot} from 'app/utils/saveGame';
