+++ app/scenes/scenes.ts
@@ -1,0 +1,3 @@
+export * from 'app/scenes/field/fieldScene';
+export * from 'app/scenes/intro/introScene';
+export * from 'app/scenes/title/titleScene';
